Handle ELF binary with no program segments
Do not return False for NX and No for RELRO when there is no program
segments in the ELF binary file (e.g. kernel module).

This is inspired from
https://github.com/slimm609/checksec/commit/29aea688e10b5c16d1cf60e3968f170b2147c150

For RELRO, a new NA value is added
For NX, True is returned to avoid changing NX type from boolean to
string

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>